requirements.txt
requests
boto3
schedule
python-dotenv
